Safari-611.1.10/LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/none-sw-from-none.https-expected.txt

CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html due to access control checks.

FAIL setting up promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')"
PASS making a same-origin request for CORP: same-origin
PASS making a same-origin request for no CORP
PASS making a same-origin request for CORP: cross-origin
FAIL making a cross-origin request for CORP: same-origin assert_unreached: Should have rejected: undefined Reached unreachable code
PASS making a cross-origin request for no CORP
PASS making a cross-origin request for CORP: cross-origin
FAIL making a cross-origin request for CORP: same-origin [PASS THROUGH] assert_unreached: Should have rejected: undefined Reached unreachable code
PASS making a cross-origin request for no CORP [PASS THROUGH]
PASS making a cross-origin request for CORP: cross-origin [PASS THROUGH]
PASS making a cross-origin request with CORS without ACAO
PASS making a cross-origin request with CORS
